feat(): add `ignoresSafeArea` parameter to `ViewLayout`
