+++ Sources/JSToast/Toast/Layout.swift
@@ -69,58 +69,6 @@ public extension Layout {
     }
 }
 
-public struct ViewLayout {
-    // MARK: - Property
-    let layout: (UIView?) -> any Layout
-    
-    // MARK: - Initializer
-    init(_ layout: @escaping (UIView?) -> any Layout) {
-        self.layout = layout
-    }
-    
-    // MARK: - Public
-    public static func inside(
-        _ anchor: Anchor,
-        offset: CGFloat = 0
-    ) -> ViewLayout {
-        ViewLayout { target in
-            .inside(anchor, of: target, offset: offset, ignoresSafeArea: true)
-        }
-    }
-    
-    public static func outside(
-        _ anchor: Anchor,
-        offset: CGFloat = 0
-    ) -> ViewLayout {
-        ViewLayout { target in
-            .outside(anchor, of: target, offset: offset)
-        }
-    }
-    
-    public static func center(
-        _ axis: Axis,
-        offset: CGFloat = 0
-    ) -> ViewLayout {
-        ViewLayout { target in
-            .center(axis, of: target, offset: offset, ignoresSafeArea: true)
-        }
-    }
-    
-    public static func width(_ width: CGFloat) -> ViewLayout {
-        ViewLayout { _ in
-            .width(width)
-        }
-    }
-    
-    public static func height(_ height: CGFloat) -> ViewLayout {
-        ViewLayout { _ in
-            .height(height)
-        }
-    }
-    
-    // MARK: - Private
-}
-
 // MARK: - Inside Layout
 public struct InsideLayout: Layout {
     // MARK: - Property
